Fix type checking

+++ levelling.ts
@@ -9,24 +9,24 @@ export async function handleLevelling(db: Database, message: Message) {
   // Mee6 formular taken from https://wiki.cakey.bot/en/feature/leveling and/or https://github.com/Mee6/Mee6-documentation/blob/master/docs/levels_xp.md
   const min = 15;
   const max = 25;
-  const gainedXp = BigInt(Math.floor(randomIntFromInterval(min, max)));
+  const gainedXp = Math.floor(randomIntFromInterval(min, max));
   const databaseTable = (await db.getLevelling(message.author.id))[0];
   // at max. once per minute allowed to avoid spamming
   if (
     databaseTable &&
-    databaseTable.lastMessageTimestamp + 60n * 1000n >= message.createdTimestamp
+    databaseTable.lastMessageTimestamp + 60 * 1000 >= message.createdTimestamp
   ) {
     return;
   }
-  const xp = (databaseTable?.xp ?? 0n) + gainedXp;
+  const xp = (databaseTable?.xp ?? 0) + gainedXp;
   let currentLvl = databaseTable?.lvl ?? 0;
-  let nextLvlXp = !databaseTable || databaseTable.nextLvlXp == 0n
-    ? 5n * 1n ** 2n + 50n * 1n + 100n - xp
+  let nextLvlXp = !databaseTable || databaseTable.nextLvlXp == 0
+    ? 5 * 1 ** 2 + 50 * 1 + 100 - xp
     : databaseTable.nextLvlXp;
   if (xp >= nextLvlXp) {
     // Level-Up
-    currentLvl += 1n;
-    nextLvlXp = 5n * currentLvl ** 2n + 50n * currentLvl + 100n - xp;
+    currentLvl += 1;
+    nextLvlXp = 5 * currentLvl ** 2 + 50 * currentLvl + 100 - xp;
     const levelupMessage = Deno.env.get("LEVELUP_MESSAGE")?.replace(
       "{player}",
       `<@${message.author.id}>`,
@@ -66,7 +66,7 @@ export async function handleLevelling(db: Database, message: Message) {
   }
   db.updateLevelling(
     message.author.id,
-    BigInt(message.createdTimestamp), /*milliseconds*/
+    message.createdTimestamp, /*milliseconds*/
     xp,
     currentLvl,
     nextLvlXp,
